+++ CHANGELOG.md
@@ -13,6 +13,9 @@ __Special thanks to @gamzeozgul for contributing to this release.__
   function. These options allow geometries to compare as equal, even when their
   Z and M values differ slightly.
 
+- Adds UTM projection support to the `carto` package. This allows projecting
+  geometries between angular (lon/lat) coordinates and UTM (x/y) coordinates.
+
 - Adds new `FlipCoordinates` methods to each geometry type. This method creates
   a new geometry with the X and Y coordinates interchanged.